% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_econ_sum.R
\name{get_econ_sum}
\alias{get_econ_sum}
\title{Get Economic Summary}
\usage{
get_econ_sum(expo_type = "World Expo")
}
\arguments{
\item{expo_type}{Type of expo: World Expo or Specialised Expo}
}
\value{
A dataframe of economic factors
}
\description{
Returns a summary of economic factors
}
