% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_themes.R
\name{get_top_themes}
\alias{get_top_themes}
\title{Get Top Themes}
\usage{
get_top_themes(expo_type = "World Expo")
}
\arguments{
\item{expo_type}{Type of expo: World Expo or Specialised Expo}
}
\value{
A dataframe of top themes
}
\description{
Returns the top themes for World or Specialised Expos
}
